only need to refactor and comment implementation

+++ backend-java/FullstackJava/comment-service/src/main/java/be/pxl/services/controller/CommentController.java
@@ -13,25 +13,45 @@ public class CommentController {
     private final CommentService commentService;
 
     @PostMapping("/{postId}")
-    public ResponseEntity<?> saveComment(@PathVariable Long postId, @RequestBody CommentRequest commentRequest) {
+    public ResponseEntity<?> saveComment(@RequestHeader(value = "role") String role, @PathVariable Long postId, @RequestBody CommentRequest commentRequest) {
+        if (role == null) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         commentService.addComment(postId, commentRequest);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/{commentId}")
-    public ResponseEntity<?> deleteComment(@PathVariable Long commentId) {
+    public ResponseEntity<?> deleteComment(@RequestHeader(value = "role") String role, @PathVariable Long commentId) {
+        if (role == null) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         commentService.deleteComment(commentId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PutMapping("/{commentId}")
-    public ResponseEntity<?> updateComment(@PathVariable Long commentId, @RequestBody CommentRequest commentRequest) {
+    public ResponseEntity<?> updateComment(@RequestHeader(value = "role") String role, @PathVariable Long commentId, @RequestBody CommentRequest commentRequest) {
+        if (role == null) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         commentService.updateComment(commentId, commentRequest);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @GetMapping
-    public ResponseEntity<?> getComments() {
-        return new ResponseEntity<>(commentService.getComments(), HttpStatus.OK);
+    @GetMapping("/{postid}")
+    public ResponseEntity<?> getComments(@RequestHeader(value = "role") String role, @PathVariable Long postid) {
+        if (role == null) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        return new ResponseEntity<>(commentService.getCommentsByPostId(postid), HttpStatus.OK);
+    }
+
+    @GetMapping("/{commentId}")
+    public ResponseEntity<?> getComment(@RequestHeader(value = "role") String role, @PathVariable Long commentId) {
+        if (role == null) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        return new ResponseEntity<>(commentService.getCommentById(commentId), HttpStatus.OK);
     }
 }
